@@ -1,7 +1,8 @@
+import 'package:connectivity_plus_platform_interface/connectivity_plus_platform_interface.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:path_provider_platform_interface/path_provider_platform_interface.dart';
 import 'package:plugin_platform_interface/plugin_platform_interface.dart';
-import 'package:record_platform_interface/record_platform_interface.dart';
+import 'package:record/record.dart';
 
 const String kTemporaryPath = 'temporaryPath';
 const String kApplicationSupportPath = 'applicationSupportPath';
@@ -140,4 +141,18 @@ class FakeRecordPlatform extends Fake
 
   @override
   Future<void> dispose(String recorderId) async {}
+}
+
+class FakeConnectivityPlatform extends Fake
+    with MockPlatformInterfaceMixin
+    implements ConnectivityPlatform {
+  @override
+  Future<List<ConnectivityResult>> checkConnectivity() {
+    return Future.value([ConnectivityResult.wifi]);
+  }
+
+  @override
+  Stream<List<ConnectivityResult>> get onConnectivityChanged {
+    return Stream.value([ConnectivityResult.wifi]);
+  }
 }
